+++ docs/src/State.fs
@@ -25,6 +25,7 @@ let pageParser : Parser<Page -> Page, Page> =
             map (Component Level) (s "components" </> s "level")
             map (Component Breadcrumb) (s "components" </> s "breadcrumb")
             map (Component Card) (s "components" </> s "card")
+            map (Component Media) (s "components" </> s "media")
             map Home top ]
 
 let urlUpdate (result : Option<Page>) model =
@@ -52,7 +53,8 @@ let init result =
         { Panel = Components.Panel.State.init ()
           Level = Components.Level.State.init ()
           Breadcrumb = Components.Breadcrumb.State.init ()
-          Card = Components.Card.State.init () }
+          Card = Components.Card.State.init ()
+          Media = Components.Media.State.init () }
 
     let (model, cmd) =
         urlUpdate result { CurrentPage = Home
@@ -142,3 +144,8 @@ let update msg model =
         let (card, cardMsg) = Components.Card.State.update msg model.Components.Card
         { model with Components =
                         { model.Components with Card = card } }, Cmd.map CardMsg cardMsg
+
+    | MediaMsg msg ->
+        let (media, mediaMsg) = Components.Media.State.update msg model.Components.Media
+        { model with Components =
+                        { model.Components with Media = media } }, Cmd.map MediaMsg mediaMsg