+++ docs/src/Types.fs
@@ -19,6 +19,7 @@ type Msg =
     | LevelMsg of Components.Level.Types.Msg
     | BreadcrumbMsg of Components.Breadcrumb.Types.Msg
     | CardMsg of Components.Card.Types.Msg
+    | MediaMsg of Components.Media.Types.Msg
 
 type ElementsModel =
     { Box : Elements.Box.Types.Model
@@ -36,7 +37,8 @@ type ComponentsModel =
     { Panel : Components.Panel.Types.Model
       Level : Components.Level.Types.Model
       Breadcrumb : Components.Breadcrumb.Types.Model
-      Card : Components.Card.Types.Model }
+      Card : Components.Card.Types.Model
+      Media : Components.Media.Types.Model }
 
 type Model =
     { CurrentPage : Page