+++ docs/src/App.fs
@@ -59,7 +59,8 @@ let menu currentPage =
               menuItem "Level" (Component Level) currentPage
               menuItem "Breadcrumb" (Component Breadcrumb) currentPage
               menuItem "Card" (Component Card) currentPage
-              menuItem "Media" (Component Components.Media) currentPage ] ]
+              menuItem "Media" (Component Components.Media) currentPage
+              menuItem "Menu" (Component Components.Menu) currentPage ] ]
 
 let header =
     div [ ClassName "hero is-primary" ]
@@ -92,6 +93,7 @@ let root model dispatch =
             | Breadcrumb -> Components.Breadcrumb.View.root model.Components.Breadcrumb (BreadcrumbMsg >> dispatch)
             | Card -> Components.Card.View.root model.Components.Card (CardMsg >> dispatch)
             | Components.Media -> Components.Media.View.root model.Components.Media (MediaMsg >> dispatch)
+            | Menu -> Components.Menu.View.root model.Components.Menu (MenuMsg >> dispatch)
 
     div []
         [ div [ ClassName "navbar-bg" ]